Update fr.json
Hi, the french translation is meaningless. I made some changes to the file but it needs changes on templates and other translations to work out.
Exemple : English "Add Liquidy" = {$_('general.add')} {$_('general.liquidity')} but you cant do that in french. Its "Ajouter de la liquidité".
So you need to create a full translation for this title and not stitch two words together.