PIE-627 Add Compounding manual on /#/Learn
Added a Block at the Learn About PieDAO, being this the last block. URL and content are, in my opinion ok. Missing an image to represent the compounding process from Slice to veDough.

Additionally, trimmed the 'What are PieVaults' description to fit the element as follows:
from: PieVaults are the DAO’s latest product range, unlocking active yield-generating strategies for each underlying asset in the index. Lending, native staking, and meta-governance are all possible. Yearn Ecosystem Pie ($YPIE) is the first PieVault, giving complete exposure to the Yearn Finance ecosystem.
To: PieVaults are the DAO’s latest product range, unlocking active yield-generating strategies for each underlying asset in the index. Lending, native staking, and meta-governance are all possible with them.